pyproject.toml
[project]
name = "ai-telegram-assistant"
version = "1.0.0"
description = "Telegram bot with OpenAI GPT-4 integration, conversation memory, and multi-user support"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.7",
    "openai>=1.12.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.1.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "fakeredis>=2.21.0",
    "ruff>=0.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]
